* commit 'a58b94b00e7485cfdfd194f9f8d95cf0c617e718':
  Removing compilation warning from ODP codebase

+++ example/lpm-ipfwd/lpmlib/ip/ip_output.c
@@ -97,7 +97,7 @@ int32_t ip_send_multi(odp_packet_t pkt[],
 	}
 
 	ll_cache = neigh->ll_cache;
-	if (odp_likely(ll_cache)) {
+	if (odp_likely((long)ll_cache)) {
 		struct ether_hdr *ll_hdr;
 
 		if (odp_pktout_queue(neigh->pktio, &pktout, 1) != 1) {
